Log to short code

+++ plugins/cb_data.py
@@ -48,72 +48,14 @@ async def doc(bot,update):
      file_path = f"downloads/{new_filename}"
      message = update.message.reply_to_message
      file = message.document or message.video or message.audio
-     value = 2099999999
-     if value < file.file_size:
-     	ms = await update.message.edit("``` Trying To Download...```")
-     	c_time = time.time()
-     	try:
-     		path = await bot.download_media(message = file, progress=progress_for_pyrogram,progress_args=( "``` Trying To Download...```",  ms, c_time   ))
-     		
-     	except Exception as e:
-     		await ms.edit(e)
-     		return
-     	splitpath = path.split("/downloads/")
-     	dow_file_name = splitpath[1]
-     	old_file_name =f"downloads/{dow_file_name}"
-     	os.rename(old_file_name,file_path)
-     	user_id = int(update.message.chat.id)
-     	thumb = find(user_id)
-     	if thumb:
-     		ph_path = await bot.download_media(thumb)
-     		Image.open(ph_path).convert("RGB").save(ph_path)
-     		img = Image.open(ph_path)
-     		img.resize((320, 320))
-     		img.save(ph_path, "JPEG")
-     		c_time = time.time()
-     		await ms.edit("```Trying To Upload```")
-     		c_time = time.time()
-     		try:
-     			filw = await app.send_document(log_channel,document = file_path,thumb=ph_path,caption = f"**{new_filename}**",progress=progress_for_pyrogram,progress_args=( "```Trying To Uploading```",  ms, c_time   ))
-     			from_chat = filw.chat.id
-     			mg_id = filw.id
-     			time.sleep(2)
-     			await bot.copy_message(update.from_user.id,from_chat,mg_id)
-     			    			
-     			await ms.delete()
-     			os.remove(file_path)
-     			os.remove(ph_path)
-     		except Exception as e:
-     			await ms.edit(e)
-     			os.remove(file_path)
-     			os.remove(ph_path)
-     			return
-     	else:
-     		
-     		await ms.edit("```Trying To Upload```")
-     		c_time = time.time()
-     		try:
-     			filw = await app.send_document(log_channel,document = file_path,caption = f"**{new_filename}**",progress=progress_for_pyrogram,progress_args=( "```Trying To Uploading```",  ms, c_time   ))
-     			from_chat = filw.chat.id
-     			mg_id = filw.id
-     			time.sleep(2)
-     			await bot.copy_message(update.from_user.id,from_chat,mg_id)
-     			
-     			await ms.delete()
-     			os.remove(file_path)
-     		except Exception as e:
-     			await ms.edit(e)
-     			os.remove(file_path)
-     			return
-     		
      ms = await update.message.edit("``` Trying To Download...```")
      c_time = time.time()
      try:
-     	path = await bot.download_media(message = file, progress=progress_for_pyrogram,progress_args=( "``` Trying To Download...```",  ms, c_time   ))
+     		path = await bot.download_media(message = file, progress=progress_for_pyrogram,progress_args=( "``` Trying To Download...```",  ms, c_time   ))
+     		
      except Exception as e:
-     	await ms.edit(e)
-     	return
-     	
+     		await ms.edit(e)
+     		return
      splitpath = path.split("/downloads/")
      dow_file_name = splitpath[1]
      old_file_name =f"downloads/{dow_file_name}"
@@ -127,122 +69,71 @@ async def doc(bot,update):
      		img.resize((320, 320))
      		img.save(ph_path, "JPEG")
      		c_time = time.time()
-     		await ms.edit("```Trying To Upload```")
-     		c_time = time.time()
-     		try:
-     			await bot.send_document(update.message.chat.id,document = file_path,thumb=ph_path,caption = f"**{new_filename}**",progress=progress_for_pyrogram,progress_args=( "```Trying To Uploading```",  ms, c_time   ))
-     			await ms.delete()
-     			os.remove(file_path)
-     			os.remove(ph_path)
-     		except Exception as e:
-     			await ms.edit(e)
-     			os.remove(file_path)
-     			os.remove(ph_path)
-     			return
-     			     		     		
+     		
+     else:
+     		ph_path = None
+     value = 2147483648
+     if value < file.file_size:
+         await ms.edit("```Trying To Upload```")
+         try:
+             filw = await app.send_document(log_channel,document = file_path,thumb=ph_path,caption = f"**{new_filename}**",progress=progress_for_pyrogram,progress_args=( "```Trying To Uploading```",  ms, c_time   ))
+             from_chat = filw.chat.id
+             mg_id = filw.id
+             time.sleep(2)
+             await bot.copy_message(update.from_user.id,from_chat,mg_id)
+             await ms.delete()
+             os.remove(file_path)
+             try:
+                 os.remove(ph_path)
+             except:
+                 pass
+         except Exception as e:
+             await ms.edit(e)
+             os.remove(file_path)
+             try:
+                 os.remove(ph_path)
+             except:
+                 return
      else:
      		await ms.edit("```Trying To Upload```")
      		c_time = time.time()
      		try:
-     			await bot.send_document(update.message.chat.id,document = file_path,caption = f"**{new_filename}**",progress=progress_for_pyrogram,progress_args=( "```Trying To Uploading```",  ms, c_time   ))
+     			await bot.send_document(update.from_user.id,document = file_path,thumb=ph_path,caption = f"**{new_filename}**",progress=progress_for_pyrogram,progress_args=( "```Trying To Uploading```",  ms, c_time   ))			
      			await ms.delete()
      			os.remove(file_path)
      		except Exception as e:
      			await ms.edit(e)
      			os.remove(file_path)
-     			return
+     			return 
      			     		   		
-     		
+   		
+     		     		     		
 @Client.on_callback_query(filters.regex("vid"))
 async def vid(bot,update):
-     new_name = update.message.text
+   new_name = update.message.text
      name = new_name.split(":-")
      new_filename = name[1]
      file_path = f"downloads/{new_filename}"
      message = update.message.reply_to_message
      file = message.document or message.video or message.audio
-     value = 2099999999
-     if value < file.file_size:
-     	ms = await update.message.edit("``` Trying To Download...```")
-     	c_time = time.time()
-     	try:
-     		path = await bot.download_media(message = file, progress=progress_for_pyrogram,progress_args=( "``` Trying To Download...```",  ms, c_time   ))
-     		
-     	except Exception as e:
-     		await ms.edit(e)
-     		return
-     	splitpath = path.split("/downloads/")
-     	dow_file_name = splitpath[1]
-     	old_file_name =f"downloads/{dow_file_name}"
-     	os.rename(old_file_name,file_path)
-     	user_id = int(update.message.chat.id)
-     	thumb = find(user_id)
-     	duration = 0
-     	metadata = extractMetadata(createParser(file_path))
-     	if metadata.has("duration"):
-     		duration = metadata.get('duration').seconds
-     	if thumb:
-     		ph_path = await bot.download_media(thumb)
-     		Image.open(ph_path).convert("RGB").save(ph_path)
-     		img = Image.open(ph_path)
-     		img.resize((320, 320))
-     		img.save(ph_path, "JPEG")
-     		c_time = time.time()
-     		
-     		await ms.edit("```Trying To Upload```")
-     		c_time = time.time()
-     		try:
-     			filw = await app.send_video(log_channel,video= file_path,thumb=ph_path,caption = f"**{new_filename}**",duration =duration,progress=progress_for_pyrogram,progress_args=( "```Trying To Uploading```",  ms, c_time   ))
-     			from_chat = filw.chat.id
-     			mg_id = filw.id
-     			time.sleep(2)
-     			await bot.copy_message(update.from_user.id,from_chat,mg_id)
-     			await ms.delete()
-     			os.remove(file_path)
-     			os.remove(ph_path)
-     		except Exception as e:
-     			await ms.edit(e)
-     			os.remove(file_path)
-     			os.remove(ph_path)
-     			return
-     	else:
-     		
-     		await ms.edit("```Trying To Upload```")
-     		c_time = time.time()
-     		try:
-     			filw = await app.send_video(log_channel,video = file_path,caption = f"**{new_filename}**",duration = duration,progress=progress_for_pyrogram,progress_args=( "```Trying To Uploading```",  ms, c_time   ))
-     			from_chat = filw.chat.id
-     			mg_id = filw.id
-     			time.sleep(2)
-     			await bot.copy_message(update.from_user.id,from_chat,mg_id)
-     			
-     			await ms.delete()
-     			os.remove(file_path)
-     		except Exception as e:
-     			await ms.edit(e)
-     			os.remove(file_path)
-     			return
-     			
-     	     	
-     	
      ms = await update.message.edit("``` Trying To Download...```")
      c_time = time.time()
      try:
-     	path = await bot.download_media(message = file, progress=progress_for_pyrogram,progress_args=( "``` Trying To Download...```",  ms, c_time   ))
+     		path = await bot.download_media(message = file, progress=progress_for_pyrogram,progress_args=( "``` Trying To Download...```",  ms, c_time   ))
+     		
      except Exception as e:
-     	await ms.edit(e)
-     	return
-     
+     		await ms.edit(e)
+     		return
      splitpath = path.split("/downloads/")
      dow_file_name = splitpath[1]
      old_file_name =f"downloads/{dow_file_name}"
      os.rename(old_file_name,file_path)
-     duration = 0
-     metadata = extractMetadata(createParser(file_path))
-     if metadata.has("duration"):
-     		duration = metadata.get('duration').seconds
      user_id = int(update.message.chat.id)
      thumb = find(user_id)
+     duration = 0     
+     metadata = extractMetadata(createParser(file_path))
+     if metadata.has("duration"):
+         duration = metadata.get('duration').seconds
      if thumb:
      		ph_path = await bot.download_media(thumb)
      		Image.open(ph_path).convert("RGB").save(ph_path)
@@ -250,28 +141,43 @@ async def vid(bot,update):
      		img.resize((320, 320))
      		img.save(ph_path, "JPEG")
      		c_time = time.time()
-     		await ms.edit("```Trying To Upload```")
-     		c_time = time.time()
-     		try:
-     			filw = await bot.send_video(update.message.chat.id,video = file_path,caption = f"**{new_filename}**",thumb=ph_path,duration =duration, progress=progress_for_pyrogram,progress_args=( "```Trying To Uploading```",  ms, c_time   ))
-     			
-     			await ms.delete()
-     			os.remove(file_path)   				
-     		except Exception as e:
-     				await ms.edit(e)
-     				os.remove(file_path)
-     				
+     		
+     else:
+     		ph_path = None
+     value = 2147483648
+     if value < file.file_size:
+         await ms.edit("```Trying To Upload```")
+         try:
+             filw = await app.send_video(log_channel,document = file_path,thumb=ph_path,caption = f"**{new_filename}**",progress=progress_for_pyrogram,progress_args=( "```Trying To Uploading```",  ms, c_time   ))
+             from_chat = filw.chat.id
+             mg_id = filw.id
+             time.sleep(2)
+             await bot.copy_message(update.from_user.id,from_chat,mg_id)
+             await ms.delete()
+             os.remove(file_path)
+             try:
+                 os.remove(ph_path)
+             except:
+                 pass
+         except Exception as e:
+             await ms.edit(e)
+             os.remove(file_path)
+             try:
+                 os.remove(ph_path)
+             except:
+                 return
      else:
      		await ms.edit("```Trying To Upload```")
      		c_time = time.time()
      		try:
-     			await bot.send_video(update.message.chat.id,video = file_path,caption = f"**{new_filename}**",duration = duration, progress=progress_for_pyrogram,progress_args=( "```Trying To Uploading```",  ms, c_time   ))
+     			await bot.send_video(update.from_user.id,document = file_path,thumb=ph_path,caption = f"**{new_filename}**",progress=progress_for_pyrogram,progress_args=( "```Trying To Uploading```",  ms, c_time   ))			
      			await ms.delete()
      			os.remove(file_path)
      		except Exception as e:
      			await ms.edit(e)
      			os.remove(file_path)
      			return
+   
    
      			     		     		
 @Client.on_callback_query(filters.regex("aud"))
